Add product service, handle create product in controller

+++ shopapp/src/main/java/com/example/shopapp/dtos/ProductDTO.java
@@ -22,14 +22,14 @@ public class ProductDTO {
 
     @Min(value = 0, message = "Price must be greater than or equal to 0")
     @Max(value = 10000000, message = "Price must be less than or equal to 10,000,000")
-    private double price;
+    private Float price;
 
     private String thumbnail;
 
     private String description;
 
     @JsonProperty("category_id")
-    private long categoryId;
+    private Long categoryId;
 
     private List<MultipartFile> files;
 }
